demux: Enable all programs in a stream when requested.
(0, NULL) in the parameters of DEMUX_SET_GROUP means that all programs
in the stream should be enabled immediately, but they have remained
disabled until the next version of PAT arrived. This patch fix the
issue.

The parameters, (0, NULL), are used when --sout-all --sout=XXX command
line options are specified, so I tested with them.

Signed-off-by: HIRANO Takahito <hiranotaka@zng.info>